Reset index.html and comment out style.css
After working on laying out the html and css for the navbar and header
was running into a number of issues with flexbox. Reviewed items in
chapters on flexbox in The Odin Project and specifically Brad Traversy's
Crash Course on this topic and decided to reset my html using the idea
of containers that he uses. I tried to set up containers, but was adding
more code than I needed to do it.